requirements.txt
streamlit
pandas
numpy
requests
scikit-surprise
